swig: bind all parameters
