+++ libraries/tuttle/src/tuttle/host/attribute/ParamCustom.i
@@ -1,0 +1,20 @@
+%include <tuttle/host/global.i>
+
+%include <tuttle/host/attribute/Param.i>
+%include <tuttle/host/attribute/AnimatedParamType.i>
+
+%include <tuttle/host/ofx/attribute/OfxhParam.i>
+%include <tuttle/host/ofx/attribute/OfxhMultiDimParam.i>
+%include <tuttle/host/ofx/attribute/OfxhParamAccessor.i>
+%include <tuttle/host/ofx/attribute/OfxhParamDescriptor.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParam.hpp>
+#include <tuttle/host/ofx/attribute/OfxhParamAccessor.hpp>
+#include <tuttle/host/ofx/attribute/OfxhParamDescriptor.hpp>
+
+#include <tuttle/host/attribute/ParamCustom.hpp>
+%}
+
+%include <tuttle/host/attribute/ParamCustom.hpp>
+
