+++ libraries/tuttle/src/tuttle/host/attribute/ValueInterpolator.i
@@ -1,0 +1,7 @@
+%include <tuttle/host/global.i>
+
+%{
+#include <tuttle/host/attribute/ValueInterpolator.hpp>
+%}
+
+%include <tuttle/host/attribute/ValueInterpolator.hpp>
