+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhKeyframeParam.i
@@ -1,0 +1,7 @@
+%include <tuttle/host/global.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhKeyframeParam.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhKeyframeParam.hpp>
