+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamCustom.i
@@ -1,0 +1,9 @@
+%include <tuttle/host/global.i>
+%include <tuttle/host/ofx/attribute/OfxhParamString.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamCustom.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamCustom.hpp>
+
