+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamGroup.i
@@ -1,0 +1,10 @@
+%include <tuttle/host/global.i>
+%include <tuttle/host/ofx/attribute/OfxhParam.i>
+%include <tuttle/host/ofx/attribute/OfxhParamSet.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamGroup.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamGroup.hpp>
+
