+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamPage.i
@@ -1,0 +1,9 @@
+%include <tuttle/host/global.i>
+%include <tuttle/host/ofx/attribute/OfxhParam.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamPage.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamPage.hpp>
+
