+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamPushButton.i
@@ -1,0 +1,10 @@
+%include <tuttle/host/global.i>
+%include <tuttle/host/ofx/attribute/OfxhParam.i>
+%include <tuttle/host/ofx/attribute/OfxhKeyframeParam.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamPushButton.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamPushButton.hpp>
+
