+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamSetAccessor.i
@@ -1,0 +1,8 @@
+%include <tuttle/host/global.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamSetAccessor.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamSetAccessor.hpp>
+
