+++ libraries/tuttle/src/tuttle/host/ofx/attribute/OfxhParamSetDescriptor.i
@@ -1,0 +1,9 @@
+%include <tuttle/host/global.i>
+%include <tuttle/host/ofx/attribute/OfxhParamSetAccessor.i>
+
+%{
+#include <tuttle/host/ofx/attribute/OfxhParamSetDescriptor.hpp>
+%}
+
+%include <tuttle/host/ofx/attribute/OfxhParamSetDescriptor.hpp>
+
