--- libraries/tuttle/src/tuttle/host/ofx/attribute/allParams.i
@@ -1,30 +1,0 @@
-%include <tuttle/host/global.i>
-
-%include <tuttle/host/ofx/attribute/OfxhParam.i>
-%include <tuttle/host/ofx/attribute/OfxhParamAccessor.i>
-%include <tuttle/host/ofx/attribute/OfxhParamDescriptor.i>
-
-%{
-#include <tuttle/host/ofx/attribute/OfxhParam.hpp>
-
-#include <tuttle/host/ofx/attribute/OfxhParamBoolean.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamChoice.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamDouble.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamGroup.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamInteger.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamPage.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamPushButton.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamString.hpp>
-#include <tuttle/host/ofx/attribute/OfxhParamCustom.hpp>
-%}
-
-%include <tuttle/host/ofx/attribute/OfxhParamBoolean.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamChoice.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamDouble.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamGroup.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamInteger.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamPage.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamPushButton.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamString.hpp>
-%include <tuttle/host/ofx/attribute/OfxhParamCustom.hpp>
-
